Prints the words that contain the letters you type in
Enter a bunch of letters and the program will traverse the whole
dictionary to find the words that contain all the given letters.
Couldn't figure out how to deal with duplicates unfortunately.

letterValidator's written off the original wordValidator, so if Casey
made any changes, they're not in letterValidator, sorry.

Replace the "print valid result" section in every search subroutine with
"add buffer contents to Maggy's list" and it should work.